refactor(model): move resolveTextAnchor into the shared layer
A second reader has appeared for it. The annotation layer's in-place
projection asks where a passage sits in a body that has since been
edited; a proposed replacement passage (ADR-0029 decision 6) has to ask
exactly the same question before it can say what adopting it does. The
resolver lived in apps/web, which a shared-layer package cannot reach,
so the second reader's only option was a second implementation — and two
answers disagree exactly where it matters: a passage one calls lost
while the other rewrites it.

Its own comment already argued this for the two keeper pages ("One
function so the two keeper pages cannot answer this differently — the
gap that let the rail exist on one page and not the other was exactly
this shape, a decision made twice"). This is the same argument one layer
up.

What moved is the pure half: resolveTextAnchor, LivePassage,
ResolvedTextAnchor. What stayed is the thread-shaped half —
markdownAnchorResolver and missingThreadMarks are about the rail and
about a document's marks, neither of which model knows. apps/web's
module re-exports the moved names so no call site changes.

The property test moves with the code it tests; model's fast-check
helper is character-identical to the one it used. model-node now covers
the resolution order directly (5 examples + the property) rather than
only through a composition root.
