+++ packages/model/src/index.ts
@@ -14,4 +14,5 @@ export * from './markdown.js'
 export * from './proposal.js'
 export * from './proposal-apply.js'
 export * from './spatial.js'
+export * from './text-anchor.js'
 export * from './trust.js'
