+++ packages/model/src/text-anchor.property.test.ts
@@ -13,7 +13,7 @@
  * which would be building the oracle out of the code under test.
  */
 import { describe, expect, it } from 'vitest'
-import { fc, fcTest, withDefaults } from '../test-utils/fast-check.js'
+import { fc, fcTest, withDefaults } from './test-utils/fast-check.js'
 import { resolveTextAnchor, type TextAnchor } from './text-anchor.js'
 
 /** Ordinary prose, and never the passage's own text — see `passage`. */
